requirements.txt
augmax==0.3.4
dm-tree==0.1.8
einops==0.8.0
equinox==0.11.8
flatbuffers==24.3.25
flax==0.1.2
fsspec[gcs]==2024.6.0
gym-aloha==0.1.1
imageio==2.36.1
jax[cuda12]==0.5.3
jaxtyping==0.2.36
ml_collections==1.0.0
numpy==1.22.4
numpydantic==1.6.6
opencv-Python==4.10.0.84
openpi-client
orbax-checkpoint==0.11.13
pillow==11.0.0
s3fs==2024.9.0
sentencepiece==0.2.0
torch==2.7.0
tqdm-loggable==0.2
typing-extensions==4.12.2
tyro==0.9.5
wandb==0.19.1
boto3==1.35.7
types-boto3[boto3,s3]==1.35.7
filelock==3.16.1
beartype==0.19.0
treescope==0.1.7
transformers==4.48.1
rich==14.0.0
polars==1.30.0
pytest==8.3.4
ruff==0.8.6
pre-commit==4.0.1
ipykernel==6.29.5
ipywidgets==8.1.5
matplotlib==3.10.0
pynvml==12.0.0
ml-dtypes==0.4.1
tensorstore==0.1.74